Add prop to disable file extension checking
