+++ src/components/modal-container.js
@@ -61,23 +61,7 @@ import {
 } from 'constants/default-settings';
 import {EXPORT_MAP_FORMATS} from 'constants/default-settings';
 import KeyEvent from 'constants/keyevent';
-
-// File.type is not reliable if the OS does not have a
-// registered mapping for the extension.
-// NOTE: Shapefiles must be in a compressed format since
-// it requires multiple files to be present.
-const DEFAULT_FILE_EXTENSIONS = ['csv', 'json', 'geojson'];
-const DEFAULT_FILE_FORMATS = ['CSV', 'Json', 'GeoJSON'];
-
-const getFileFormatNames = createSelector(
-  state => state.loaders,
-  loaders => [...DEFAULT_FILE_FORMATS, ...loaders.map(loader => loader.name)]
-);
-
-const getFileExtensions = createSelector(
-  state => state.loaders,
-  loaders => [...DEFAULT_FILE_EXTENSIONS, ...loaders.flatMap(loader => loader.extensions)]
-);
+import {getFileFormatNames, getFileExtensions} from '../reducers/vis-state-selectors';
 
 const DataTableModalStyle = css`
   top: 80px;
