+++ src/components/modals/load-data-modal.d.ts
@@ -16,6 +16,8 @@ export type LoadDataModalProps = {
   fileFormatNames: string[];
   /** A list of typically 3 letter extensions (without '.') for file matching */
   fileExtensions: string[];
+  /** Set to true if app wants to do its own file filtering */
+  disableExtensionFilter?: boolean;
 };
 
 export function LoadDataModalFactory(
